Make the notification a data notification
You have much more control over a data notification. E.g. this way you can localize the text for the user. Also this way you do not need to go through the launch activity but can directly launch the correct action.
It was this way before - but with the unfortunate migration to javascript it became a message notification.